Support long long problem sizes.

+++ 01_map/kernel.cu
@@ -2,22 +2,22 @@
 #include "../utils/utils.h"
 #include "../utils/tracer.h"
 
-__global__ void map_kernel(const float* a, const float* b, float* c, int size) {
+__global__ void map_kernel(const float* a, const float* b, float* c, long long size) {
   /*
-  int lane = blockIdx.x * blockDim.x + threadIdx.x;
-  int stride = blockDim.x * gridDim.x;
+  long long lane = blockIdx.x * blockDim.x + threadIdx.x;
+  long long stride = blockDim.x * gridDim.x;
   for (;lane < size; lane += stride) {
     c[lane] = a[lane] + b[lane];
   }
   */
 }
 
-std::vector<LaunchConfig> launch_map(const float* a, const float* b, float* c, int size) {
+std::vector<LaunchConfig> launch_map(const float* a, const float* b, float* c, long long size) {
     global_tracer.trace("Entering launch_map (Student)");
     
     // TODO: Define grid and block dimensions
     int threadsPerBlock = 256;
-    int blocksPerGrid = (size + threadsPerBlock - 1) / threadsPerBlock;
+    long long blocksPerGrid = (size + threadsPerBlock - 1) / threadsPerBlock;
     
     global_tracer.trace("Launching map_kernel (Student)");
     map_kernel<<<blocksPerGrid, threadsPerBlock>>>(a, b, c, size);
