+++ 02_reduce/reference_kernel.h
@@ -4,6 +4,6 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_reference_reduce(const float* a, float* c, int size);
+std::vector<LaunchConfig> launch_reference_reduce(const float* a, float* c, long long size);
 
 #endif
