+++ 04_histogram/kernel.h
@@ -2,4 +2,4 @@
 #include <vector>
 #include "../utils/utils.h"
 
-std::vector<LaunchConfig> launch_histogram(const int* a, int* bins, int size, int num_bins);
+std::vector<LaunchConfig> launch_histogram(const int* a, int* bins, long long size, int num_bins);
